Add webhook example to do some e2e tests
